Decouple Practice Writing tab, fix toast msg z-index
